third_party/WebKit/LayoutTests/fast/dom/shadow/custom-pseudo-in-selector-api.html

<!doctype html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script src="resources/shadow-dom.js"></script>
</head>
<body>
  <div id='sandbox'></div>
  <pre id='console'></pre>
<script>
description('Test for crbug.com/273960. Vendor prefixed pseudo elements should not cause DOM exception 12 when using querySelector, querySelectorAll and webkitMatchesSelector.');

var sandbox = document.getElementById('sandbox');

sandbox.appendChild(document.createElement("input"));
shouldBeFalse('sandbox.firstChild.webkitMatchesSelector("input[type=\'search\']::-webkit-search-decoration")');
shouldBeNull('sandbox.firstChild.querySelector("input[type=\'search\']::-webkit-search-decoration")');
shouldBe('sandbox.firstChild.querySelectorAll("input[type=\'search\']::-webkit-search-decoration").length', '0');

sandbox.innerHTML = '';
</script>
</body>
</html>
